+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - "Auto-Extract Downloaded Games" option added to settings
 - Image cache and offline cache size is displayed in settings
 - Added Logout option to settings
+- Added Download Limit to settings
 
 ## 1.4.1
 ### Changes